+++ rerun_py/rerun_sdk/rerun/archetypes/transform3d.py
@@ -21,6 +21,14 @@ class Transform3D(Transform3DExt, Archetype):
     """
     **Archetype**: A transform between two 3D spaces, i.e. a pose.
 
+    All components are applied in the order they are listed here.
+    E.g. if both a 4x4 matrix with a translation and a translation vector are present,
+    the matrix is applied first, then the translation vector on top.
+
+    Each transform component can be listed multiple times, but transform tree propagation is only possible
+    if there's only one instance for each transform component.
+    TODO(#6831): write more about the exact interaction with the to be written `OutOfTreeTransform` component.
+
     Examples
     --------
     ### Variety of 3D transforms:
@@ -130,6 +138,8 @@ class Transform3D(Transform3DExt, Archetype):
         """Convenience method for calling `__attrs_init__` with all `None`s."""
         self.__attrs_init__(
             transform=None,  # type: ignore[arg-type]
+            mat3x3=None,  # type: ignore[arg-type]
+            translation=None,  # type: ignore[arg-type]
             axis_length=None,  # type: ignore[arg-type]
         )
 
@@ -145,6 +155,24 @@ class Transform3D(Transform3DExt, Archetype):
         converter=components.Transform3DBatch._required,  # type: ignore[misc]
     )
     # The transform
+    #
+    # (Docstring intentionally commented out to hide this field from the docs)
+
+    mat3x3: components.TransformMat3x3Batch | None = field(
+        metadata={"component": "optional"},
+        default=None,
+        converter=components.TransformMat3x3Batch._optional,  # type: ignore[misc]
+    )
+    # 3x3 transformation matrices.
+    #
+    # (Docstring intentionally commented out to hide this field from the docs)
+
+    translation: components.Translation3DBatch | None = field(
+        metadata={"component": "optional"},
+        default=None,
+        converter=components.Translation3DBatch._optional,  # type: ignore[misc]
+    )
+    # Translation vectors.
     #
     # (Docstring intentionally commented out to hide this field from the docs)